refactor: 전체 패키지 구조 수정
