+++ app/src/main/java/nextstep/jwp/Controller.java
@@ -11,9 +11,6 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 
 public class Controller {
-    public String index() throws IOException {
-        return view("/index.html");
-    }
 
     public String login() throws IOException {
         return view("/login.html");
@@ -41,32 +38,6 @@ public class Controller {
 
         InMemoryUserRepository.save(user);
         return redirect("/index.html");
-    }
-
-    public String css(String uri) throws IOException {
-        final URL resource = getClass().getClassLoader().getResource("static" + uri);
-        final Path path = new File(resource.getPath()).toPath();
-        String responseBody = Files.readString(path);
-
-        return String.join("\r\n",
-                "HTTP/1.1 200 OK ",
-                "Content-Type: text/css;charset=utf-8 ",
-                "Content-Length: " + responseBody.getBytes().length + " ",
-                "",
-                responseBody);
-    }
-
-    public String js(String uri) throws IOException {
-        final URL resource = getClass().getClassLoader().getResource("static" + uri);
-        final Path path = new File(resource.getPath()).toPath();
-        String responseBody = Files.readString(path);
-
-        return String.join("\r\n",
-                "HTTP/1.1 200 OK ",
-                "Content-Type: Application/javascript;charset=utf-8 ",
-                "Content-Length: " + responseBody.getBytes().length + " ",
-                "",
-                responseBody);
     }
 
     private Map<String, String> bodyMapper(String requestBody) {
