+++ app/src/main/java/nextstep/jwp/db/InMemoryUserRepository.java
@@ -1,11 +1,10 @@
 package nextstep.jwp.db;
 
 
-import nextstep.jwp.model.User;
-
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
+import nextstep.jwp.model.User;
 
 public class InMemoryUserRepository {
