+++ app/src/main/java/nextstep/jwp/http/HttpRequest.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.http;
 
 import java.util.HashMap;
 import java.util.Map;
